% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_prob_cohort.R
\name{death_prob_cohort}
\alias{death_prob_cohort}
\title{Compute Death Probabilities for a Specified Cohort at a Specified Age}
\usage{
death_prob_cohort(
  data,
  age,
  cohort,
  surv_stat = "mm2",
  DOB_CMC = "mm4",
  obs_time = "obs_time"
)
}
\arguments{
\item{data}{The relevant data frame}

\item{age}{The age for which the statistic is desired}

\item{cohort}{The cohort for which the statistic is desired}

\item{surv_stat}{String name for the column containing survival status. Requires that deceased is coded as 0, and alive is coded as 1. Defaults to 'mm2', the original DHS variable name.}

\item{DOB_CMC}{String name for column containing dates of birth in century month code. Defaults to 'mm4', the original DHS variable name.}

\item{obs_time}{String name for the column containing observation times, as can be generated by the mortDHS_obs_time function. Defaults to 'obs_time'}
}
\value{
A single numerical value
}
\description{
Calculates the probability of death for a user-specified cohort and a user-specified age.
The formula used can be found at tinyurl.com/demograsource.

For example, the question 'what was the probability of death at age 1 for the cohort born in 1995?' can be answered with this function.
This would be computed as the number of people in that cohort who died at age 1 divided by the number that reached their 1st birthday.

The function requires a data frame that contains variables for:

- date of birth in century month code

- survival status

- survival time, as can be generated by the mortDHS_obs_time function in this package.
}
\examples{
# compute probability of death at age 0 for the cohort born in 2000
death_prob_cohort(malawisib, 0, 2000)

# same example as above, with each parameter more explicitly defined
death_prob_cohort(data = malawisib, age =0, cohort = 2000,
DOB_CMC = 'mm4', surv_stat = 'mm2', obs_time = 'obs_time')
}
\seealso{
\code{\link[mortDHS]{death_prob_cohort_range}}
}
