% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_prob_year_range.R
\name{death_prob_year_range}
\alias{death_prob_year_range}
\title{Compute Death Probabilities for a Specified Year Range and a Specified Age Range}
\usage{
death_prob_year_range(
  data,
  lower_age,
  upper_age,
  lower_year,
  upper_year,
  surv_stat = "mm2",
  DOB_CMC = "mm4",
  obs_time = "obs_time"
)
}
\arguments{
\item{data}{The relevant data frame}

\item{lower_age}{The lower boundary of the age interval for which the statistic is desired}

\item{upper_age}{The upper boundary of the age interval for which the statistic is desired}

\item{lower_year}{The lower boundary of the year interval for which the statistic is desired}

\item{upper_year}{The upper boundary of the year interval for which the statistic is desired}

\item{surv_stat}{String name for the column containing survival status. Requires that deceased is coded as 0, and alive is coded as 1. Defaults to 'mm2', the original DHS variable name.}

\item{DOB_CMC}{String name for column containing dates of birth in century month code. Defaults to 'mm4', the original DHS variable name.}

\item{obs_time}{String name for the column containing observation times, as generated by the mortDHS_obs_time function. Defaults to 'obs_time'}
}
\value{
A dataframe with death probabilities. Ages are arranged along the columns, while years are arranged along the rows.
}
\description{
Calculates the probability of death for a user-specified year and age range

For example, the question 'what was the probability of death for teenagers (13-19) in the time between 1995 and the end of 1996?' can be answered with this function.
This would be an averaged value of:

-The death probability at age 13 in 1995 (as computed by the death_prob_year function)

-The death probability at age 14 in 1995

-...

-The death probability at age 19 in 1995; and then

-The death probability at age 13 in 1996

-The death probability at age 14 in 1996

-...

-The death probability at age 19 in 1996

Requires a data frame that contains the variables date of birth in century month code, survival status and observation time, as can be generated by the mortDHS_obs_time function in this package.
}
\examples{
# compute probability of death between ages 0 and 5 for cohorts born between 2000 and 2005
death_prob_year_range(malawisib, 0,5,2000,2005)

# same example as above, with each parameter more explicitly defined
death_prob_year_range(data = malawisib, lower_age = 0, upper_age = 5,
lower_year = 2000, upper_year = 2005,
DOB_CMC = 'mm4', surv_stat = 'mm2', obs_time = 'obs_time')
}
\seealso{
\code{\link[mortDHS]{death_prob_year}}
}
