% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortDHS_merge.R
\name{mortDHS_merge}
\alias{mortDHS_merge}
\title{Merge a Sibling Recode File with Select Variables from the DHS Individual Recode}
\usage{
mortDHS_merge(sib_df, respondent_df, respondent_vars = c("v008"), id = "caseid")
}
\arguments{
\item{sib_df}{The DHS sibling recode file, as generated with the mortDHS_reshape function}

\item{respondent_df}{The corresponding DHS individual respondent recode dataframe}

\item{respondent_vars}{String names of the columns in the individual recode file to be merged. Defaults to 'v008', as this is the only additional variable needed to calculate sibling mortality statistics.}

\item{id}{The name of the ID variable in the dataset, entered as a string. Defaults to 'caseid'.}
}
\value{
Returns the merged dataframe
}
\description{
This is simply a wrapper for the 'merge' function in base R with handy defaults. It allows the user to easily select which of the variables from the DHS individual recode should be included.
}
\examples{
# first generate the sibling recode dataframe
sib_recode <- mortDHS_reshape(data = malawidhs, sib_cols = c(1:5))
# then merge it with the individual recode file
merged_file <- mortDHS_merge(sib_df = sib_recode, respondent_df = malawidhs)
}
