% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortDHS_obs_time.R
\name{mortDHS_obs_time}
\alias{mortDHS_obs_time}
\title{Compute Observation Time for Survival Analysis using DHS Century Month Codes}
\usage{
mortDHS_obs_time(
  data,
  surv_stat = "mm2",
  DOB_CMC = "mm4",
  DOD_CMC = "mm8",
  interview_date_CMC = "v008"
)
}
\arguments{
\item{data}{The relevant data frame}

\item{surv_stat}{String name for the column containing survival status. Requires that deceased is coded as 0, and alive is coded as 1. Defaults to 'mm2' as in the DHS Individual Recode File.}

\item{DOB_CMC}{String name for the column containing dates of birth in century month code. Defaults to 'mm4', which is the corresponding variable name in the DHS individual recode files.}

\item{DOD_CMC}{String name for the column containing dates of death in century month code. Defaults to 'mm8', which is the corresponding variable name in the DHS individual recode files.}

\item{interview_date_CMC}{String name for the column containing dates of interview in century month code. Defaults to 'v008', which is the corresponding variable name in the DHS individual recode files.}
}
\value{
A vector of observation times
}
\description{
Generates a 'Time Under Observation' variable from DHS data, accounting for right censoring.
     For individuals who are alive at the time of the survey, observation time is their age, in months, at the start of the survey.
     For individuals who are deceased at the time of the survey, observation time is their age, in months, at the time of death.
}
\examples{
# first generate the sibling recode dataframe
sib_recode <- mortDHS_reshape(data = rwandadhs, sib_cols = c(1,2,4,8))
# then merge it with the individual recode file
merged_file <- mortDHS_merge(sib_df = sib_recode, respondent_df = rwandadhs)
# then create an observation time column in the merged data frame
merged_file$obs_time <-
   mortDHS_obs_time(merged_file, surv_stat = 'mm2' , DOB_CMC = 'mm4', DOD_CMC = 'mm8',
   interview_date_CMC = 'v008' )
}
