% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortDHS_reshape.R
\name{mortDHS_reshape}
\alias{mortDHS_reshape}
\title{Reshape a DHS individual recode file into a one-sibling-per-row file}
\usage{
mortDHS_reshape(data, sib_cols = c(1, 2, 4, 8), id = "caseid")
}
\arguments{
\item{data}{The DHS individual recode file.}

\item{sib_cols}{A vector of numbers, e.g.c(1:8), to indicate which sibling variables to reshape. Defaults to c(1,2,4,8), because "mm1", "mm2", "mm4" and "mm8" are the DHS variables needed for the mortality analysis functions in this package.}

\item{id}{The string name of the ID variable in the dataset, entered as a string. Defaults to 'caseid'.}
}
\value{
Returns a dataframe with one sibling per row
}
\description{
This function takes a DHS individual recode file, with one respondent per row, and transforms it into a sibling recode file, with one sibling per row.
     The user can choose to indicate which of the 15 sibling variables should be included in the reshaped file.
}
\examples{
mortDHS_reshape(data = malawisib, sib_cols = c(1,2,4,8))

}
