% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_lexis.R
\name{prep_for_lexis}
\alias{prep_for_lexis}
\title{Prepare Output from death_prob Function for Lexis Plot}
\usage{
prep_for_lexis(data)
}
\arguments{
\item{data}{The relevant data frame}
}
\description{
The function reshapes and rearranges the input data frame into a ‘long’ format, ready to be turned into a ggplot geom_tile visualization.
}
