% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwandasib.R
\docType{data}
\name{rwandasib}
\alias{rwandasib}
\title{Sibling Data Reshaped from 'rwandadhs' dataframe}
\format{dataframe}
\usage{
rwandasib
}
\description{
Contains data for the 144,612 siblings of individuals in the 'rwandadhs' data frame that is also included in this package.

The dataset contains the following variables:

caseid: Unique identifier for each respondent

sib_id: Unique identifier for each sibling of each respondent

v008: Date of interview in century month code

mm1: Sex of sibling (1 = male, 2 = female)

mm2: Survival status of sibling  (0 = deceased, 1 = alive )

mm4: Siblings' dates of birth in century month code
}
